auto-test/requirements.txt
pymongo
mysql-connector-python